Remove net params and assets
